+++ docs/changelog.md
@@ -6,10 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Upcoming
 
+## 5.5.1-rc.1 - 2026-07-20
+
 ### Fixed
 
 - `six-file-upload`: Fixed six-file-upload-success being emitted twice on file drop in Angular
   applications using Zone.js.
+- Updated and fixed vulnerable dependencies within the libraries
 
 ## 5.5.0 - 2026-07-03